fix: top tab bar refactor
